改进一些通用工具，附带service同时改变
